requirements.txt
streamlit
pandas
scikit-learn
matplotlib
joblib
shap
Pillow
xgboost